+++ testcases/pkcs11/getobjectsize.c
@@ -143,5 +143,5 @@ int main(int argc, char **argv)
     funcs->C_Finalize(NULL);
 
     /* make sure we return non-zero if rv is non-zero */
-    return ((rv == 0) || (rv % 256) ? rv : -1);
+    return ((rv == 0) || (rv % 256) ? (int)rv : -1);
 }
